requirements.txt
absl-py==2.1.0
astunparse==1.6.3
asyncio==3.4.3
attrs==23.2.0
bleak==0.21.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fastjsonschema==2.19.1
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
jax==0.4.26
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
nbformat==5.10.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
pandas==2.2.1
patsy==0.5.6
pillow==10.2.0
plotly==5.20.0
plotly-express==0.4.1
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyparsing==3.1.2
pytz==2024.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
statsmodels==0.14.1
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tflite==2.10.0
threadpoolctl==3.4.0
tk==0.1.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
